# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new tourist",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login tourist",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/verify-email/confirm-otp": {
      "post": {
        "operationId": "AuthController_verifyEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailOtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify email with OTP code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resend email verification OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/google/verify": {
      "post": {
        "operationId": "AuthController_googleVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify Google ID token for social login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "get": {
        "operationId": "AuthController_validateResetToken",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Validate password reset token",
        "tags": [
          "Auth"
        ]
      },
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Request password reset email",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Change password (authenticated)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/test-email": {
      "post": {
        "operationId": "AuthController_sendTestEmailPost",
        "parameters": [],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "owellrichard@gmail.com",
                    "description": "Target email address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Test email execution result"
          }
        },
        "summary": "Send test email to test SMTP or SendGrid provider configuration",
        "tags": [
          "Auth"
        ]
      },
      "get": {
        "operationId": "AuthController_sendTestEmailGet",
        "parameters": [
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Target email address (defaults to owellrichard@gmail.com)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test email execution result"
          }
        },
        "summary": "Send test email via GET request to test SMTP or SendGrid provider configuration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "AdminUsersController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isPaid",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isEmailVerified",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isBlocked",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "registrationStatus",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "in_progress",
                "complete"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List all tourist users with filters, search, pagination, and summary card stats",
        "tags": [
          "Admin — User Management"
        ]
      }
    },
    "/api/admin/users/stats": {
      "get": {
        "operationId": "AdminUsersController_getUserStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get overview summary statistics for user metrics",
        "tags": [
          "Admin — User Management"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "get": {
        "operationId": "AdminUsersController_getUserDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get full user profile details by ID",
        "tags": [
          "Admin — User Management"
        ]
      }
    },
    "/api/admin/users/{id}/block": {
      "patch": {
        "operationId": "AdminUsersController_toggleBlockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Block or unblock a user account with reason logging",
        "tags": [
          "Admin — User Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "Africa By Road API",
    "description": "Backend API for Africa By Road - Road trip community platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "EmailOtpVerifyDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "otp": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "GoogleVerifyDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "BlockUserDto": {
        "type": "object",
        "properties": {
          "isBlocked": {
            "type": "boolean",
            "description": "Whether to block (true) or unblock (false) the user",
            "example": true
          },
          "reason": {
            "type": "string",
            "description": "Reason for blocking or unblocking the user",
            "example": "Violation of community terms and conditions."
          }
        },
        "required": [
          "isBlocked"
        ]
      },
...
    }
  }
}